// src/app/model/address.model.ts
export class Address {
  ADDRESS_ID: string;
  ADDRESS_TYPE: string;
  ADDR_LINE1: string;
  ADDR_LINE2: string;
  ADDR_LINE3: string;
  ADDR_LINE4: string;
  PIN: string;
  CITY: string;
  STATE: string;
  COUNTRY: string;
  LOCATION_LONG: string;
  LOCATION_LATT: string;
  LOCATION_NAME: string;
}